Moved product analytics to EE codebase
