+++ enterprise/frontend/src/metabase-enterprise/product_analytics/ProductAnalyticsSettingsPage/SiteDetailsModal.module.css
@@ -1,0 +1,3 @@
+.CodeBlockText {
+  text-wrap-mode: wrap;
+}
