+++ enterprise/frontend/src/metabase-enterprise/product_analytics/ProductAnalyticsSettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { ProductAnalyticsSettingsPage } from "./ProductAnalyticsSettingsPage";
