+++ enterprise/frontend/src/metabase-enterprise/product_analytics/index.ts
@@ -1,0 +1,12 @@
+import { PLUGIN_PRODUCT_ANALYTICS } from "metabase/plugins";
+import { hasPremiumFeature } from "metabase-enterprise/settings";
+
+import { getProductAnalyticsAdminRoutes } from "./routes";
+
+export function initializePlugin() {
+  // TODO: add proper feature token check
+  if (hasPremiumFeature("hosting")) {
+    PLUGIN_PRODUCT_ANALYTICS.isEnabled = true;
+    PLUGIN_PRODUCT_ANALYTICS.getAdminRoutes = getProductAnalyticsAdminRoutes;
+  }
+}
