+++ enterprise/frontend/src/metabase-enterprise/product_analytics/routes.tsx
@@ -1,0 +1,9 @@
+import { Route } from "react-router";
+
+import { ProductAnalyticsSettingsPage } from "./ProductAnalyticsSettingsPage";
+
+export const getProductAnalyticsAdminRoutes = () => {
+  return (
+    <Route path="product-analytics" component={ProductAnalyticsSettingsPage} />
+  );
+};
