+++ frontend/src/metabase/admin/components/CodeTextBlock.module.css
@@ -1,0 +1,11 @@
+.InfoBlockButton {
+  position: absolute;
+  top: 0;
+  right: 0;
+  margin: 1rem;
+  cursor: pointer;
+
+  &:hover {
+    color: var(--mb-color-brand);
+  }
+}
